backend/requirements.txt
fastapi
uvicorn
anthropic
openai
Jinja2
dotenv
httpx
pytest
pyyaml
python-multipart